+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 4.3.1
++ [x] Interop: fix interop with Metro4 and jQuery, issue #1408 
 
 ### 4.3.0
 + [x] jQuery: Goodbye!!!